Implement entity validation using Apache BVal
Whenever a validation constraints fails on entity save, an
exception is raised and the entity is not persisted. I'm
catching these exceptions as generic Exception in the
GlobalErrorsHandler and returning to the client a generic
500 error code.

The proper 400 error code with the response body indicanting
the errors is not implemnted yet. That would be when validating
DTOs and getting the validation errors with the method
validator.validate().